requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.6.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
dnspython==1.15.0
feedfinder2==0.0.4
feedparser==5.2.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
jieba3k==0.35.1
lxml==4.2.1
nltk==3.3
parsel==1.4.0
Pillow==5.2.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pymongo==3.6.1
pyOpenSSL==17.5.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.13
queuelib==1.5.0
requests==2.19.1
requests-file==1.4.3
Scrapy==1.5.0
service-identity==17.0.0
six==1.11.0
tldextract==2.2.0
Twisted==17.9.0
tzlocal==1.5.1
urllib3==1.23
w3lib==1.19.0
zope.interface==4.5.0
